Update draft-ietf-ntp-roughtime.md (abstract and overview)
Extended the abstract back to two core features, and clarified the scope as limited to the on-wire protocol with the (necessary) ecosystem being unspecified (but commented on, this remains TBD).

Integrated "The Guarantee" into the Overview section. Also heavily edited its content, to be more parseable for a reader IMO, and to clarify the context of malfeasance observation vs. logging/crypto-proving vs. reporting vs. impeachment.